Revert "Updated the code to take dynamic input from user"
This reverts commit b6aeba26cd72a895cbe44fa88560385a15cd40ea.
